Section-2, creating customer layers
