+++ src/main/java/com/example/demo/DemoApplication.java
@@ -13,29 +13,8 @@ public class DemoApplication {
 		SpringApplication.run(DemoApplication.class, args);
 	}
 
-	@GetMapping
-	Customer getCustomer(){
-		return new Customer(1L,"James Bond");
 
-	}
 
-	class Customer{
-		private final Long id;
-		private final String name;
 
-		Customer(Long id, String name){
-
-			this.id = id;
-			this.name = name;
-		}
-
-		public Long getId() {
-			return id;
-		}
-
-		public String getName() {
-			return name;
-		}
-	}
 
 }
