+++ src/main/java/com/example/demo/customer/Customer.java
@@ -1,0 +1,23 @@
+package com.example.demo.customer;
+
+public class Customer {
+
+    private final Long id;
+    private final String name;
+
+
+    Customer(Long id, String name) {
+
+        this.id = id;
+        this.name = name;
+    }
+
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
