+++ src/main/java/com/example/demo/customer/CustomerController.java
@@ -1,0 +1,15 @@
+package com.example.demo.customer;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class CustomerController {
+
+    @GetMapping
+
+    Customer getCustomer(){
+        return new Customer(1L,"James Bond");
+
+    }
+}
