+++ src/main/java/com/example/demo/customer/CustomerRepository.java
@@ -1,0 +1,4 @@
+package com.example.demo.customer;
+
+public class CustomerRepository {
+}
