+++ src/main/java/com/example/demo/customer/CustomerService.java
@@ -1,0 +1,4 @@
+package com.example.demo.customer;
+
+public class CustomerService {
+}
